added page to send offer to canddiate and modified interface of candidate inboc

+++ candidate_reload_mess.py
@@ -13,17 +13,24 @@ def connect_to_database():
 def reload_messages(c_id1):
 	global cursor,conn
 	connect_to_database()
-	sql="SELECT * FROM chat where ToUserID='%s' and role='Candidate' ORDER BY Time ASC"%(c_id1)
+	sql="SELECT * FROM chat where ToUserID='%s' and role='Candidate' or role='job' ORDER BY Time ASC"%(c_id1)
 	try:
 		cursor.execute(sql)
 		results = cursor.fetchall()
 		for row in results:
+			role=row['role']
 			divid=row['ID']
-			fromuser=row["FromUser"]
 			time=row["Time"]
 			datetime=time.strftime('%H : %M')
-			text=row['Text']
-			print("""<div id="%s"  class="alert alert-info alert-dismissable fade in" ><a href="#" onclick='delete_mes("%s")' class="close" data-dismiss="alert" aria-label="close">&times;</a> %s | <strong>%s </strong> : %s </div><br/>"""%(divid,divid,datetime,fromuser,text))
+			if role=="job":
+				fromuser=row["FromUser"]
+				text=row['Text']
+				print("""<div style="margin-top:20px;background-color:white;border-left:3px solid rgba(23,139,158,1.00);border-top:2px solid rgba(23,139,158,1.00);box-shadow: 5px 5px 5px #aaaaaa;"><div class="row" style="margin-right:20px;margin-top:20px;"><div class="col-md-8"><div id="%s"  class="alert-dismissable fade in" style="margin:30px;"> %s | <strong>%s </strong> : %s </div></div><div class="col-md-4"><a href="#" onclick='delete_mes("%s")' class="close" data-dismiss="alert" aria-label="close">&times;</a></div></div></div><br/>"""%(divid,datetime,fromuser,text,divid))
+			else:
+				fromuser=row["FromUser"]
+				text=row['Text']
+				print("""<div style="margin-top:20px;background-color:white;border-left:3px solid rgba(23,139,158,1.00);border-top:2px solid rgba(23,139,158,1.00);box-shadow: 5px 5px 5px #aaaaaa;"><div class="row" style="margin-right:20px;margin-top:20px;"><div class="col-md-8"><div id="%s"  class="alert-dismissable fade in" style="margin:30px;"> %s | <strong>%s </strong> : %s </div></div><div class="col-md-4"><a href="#" onclick='delete_mes("%s")' class="close" data-dismiss="alert" aria-label="close">&times;</a></div></div></div><br/>"""%(divid,datetime,fromuser,text,divid))
+			
 	except:
 		conn.rollback()
 		print("Try again !")
@@ -43,7 +50,7 @@ def delete_message(c_id1):
 def delete_all_mess(c_id1):
 	global cursor,conn
 	connect_to_database()
-	sql="DELETE FROM chat where ToUserID='%s' and role='Candidate'"%(c_id1)
+	sql="DELETE FROM chat where ToUserID='%s' and role='Candidate' or role='job'"%(c_id1)
 	try:
 		cursor.execute(sql)
 		conn.commit()
@@ -54,7 +61,7 @@ def delete_all_mess(c_id1):
 def mess_total_count(c_id1):
 	global cursor,conn
 	connect_to_database()
-	sql="SELECT * FROM Chat where ToUserID='%s' and role='Candidate'"%(c_id1)
+	sql="SELECT * FROM Chat where ToUserID='%s' and role='Candidate' or role='job'"%(c_id1)
 	try:
 		cursor.execute(sql)
 		results=cursor.rowcount
